Provide the Names or Controls for the Prefab Models

Create a character vector of the names of the pre-fabricated (prefab) models or a list of their controls.

prefab_models()

prefab_model_controls()

Value

prefab_models: character vector of model names.

prefab_model_controls: list of model controls.

Examples

 prefab_models()
#> Warning: incomplete final line found on '/home/runner/work/_temp/Library/evercast/extdata/prefab_model_controls.yaml'
#> [1] "AutoArima" "nbGARCH"   "pGARCH"   
 prefab_model_controls()
#> Warning: incomplete final line found on '/home/runner/work/_temp/Library/evercast/extdata/prefab_model_controls.yaml'
#> $AutoArima
#> $AutoArima$metadata
#> $AutoArima$metadata$name
#> [1] "AutoArima"
#> 
#> $AutoArima$metadata$descriptor
#> [1] "classic model"
#> 
#> $AutoArima$metadata$text
#> [1] "AutoArima (Automatic Auto-Regressive Integrated Moving Average) is a flexible Auto-Regressive Integrated Moving Average (ARIMA) model fit to the data. The model is selected and  fitted using the `auto.arima` and `forecast` functions in the **forecast** package [@Hyndman2013; @Hyndman2017] within the `AutoArima()` function. Generally, ARIMA models are defined according to three model structure parameters -- the number of autoregressive terms (p), the degree of differencing (d), and the order of the moving average (q), and are represented as ARIMA(p, d, q) [@Box1970]. No seasonal models were evaluated. AutoArima is fit flexibly, such that the model parameters can vary from fit to fit."
#> 
#> 
#> $AutoArima$fun
#> [1] "auto.arima"
#> 
#> $AutoArima$args
#> $AutoArima$args$y
#> [1] "quote(abundance)"
#> 
#> 
#> $AutoArima$interpolate
#> $AutoArima$interpolate$needed
#> [1] FALSE
#> 
#> 
#> $AutoArima$datasets
#> [1] "all_total"
#> 
#> $AutoArima$response
#> $AutoArima$response$link
#> [1] "normal"
#> 
#> $AutoArima$response$type
#> [1] "distribution"
#> 
#> 
#> $AutoArima$time
#> [1] "year"
#> 
#> 
#> $nbGARCH
#> $nbGARCH$metadata
#> $nbGARCH$metadata$name
#> [1] "nbGARCH"
#> 
#> $nbGARCH$metadata$descriptor
#> [1] "classic model, requires interpolation"
#> 
#> $nbGARCH$metadata$text
#> [1] "nbGARCH (Negative Binomial Auto-Regressive Conditional Heteroskedasticity) is a generalized autoregressive conditional heteroskedasticity (GARCH) model with overdispersion (*i.e.*, a negative binomial response) fit to the interpolated data at the composite (full site and just control plots) spatial level and both the composite (community) and the articulated (species) ecological levels. The model for each species and the community total is selected and fitted using the `tsglm` function in the **tscount** package [@Liboschik2017a] within the `nbGARCH()` function. GARCH models are generalized ARMA models and are defined according to their link function, response distribution, and two model structure parameters -- the number of autoregressive terms (p) and the order of the moving average (q), and are represented as GARCH(p, q) [@Liboschik2017a]. The nbGARCH model is fit using the log link and a negative binomial response (modeled as an over-dispersed Poisson), as well as with p = 1 (first-order autoregression) and q = NULL (no moving average). The `tsglm` function in the **tscount** package [@Liboschik2017a] uses a (conditional) quasi-likelihood based approach to inference and models the overdispersion as an additional parameter in a two-step approach. This two-stage approach has only been minimally evaluated, although preliminary simulation-based studies are promising [@Liboschik2017b]."
#> 
#> 
#> $nbGARCH$fun
#> [1] "tsglm"
#> 
#> $nbGARCH$args
#> $nbGARCH$args$ts
#> [1] "abundance"
#> 
#> $nbGARCH$args$model
#> $nbGARCH$args$model$past_obs
#> [1] 1
#> 
#> $nbGARCH$args$model$past_mean
#> [1] 12
#> 
#> 
#> $nbGARCH$args$distr
#> [1] "nbinom"
#> 
#> $nbGARCH$args$link
#> [1] "log"
#> 
#> 
#> $nbGARCH$interpolate
#> $nbGARCH$interpolate$needed
#> [1] TRUE
#> 
#> $nbGARCH$interpolate$fun
#> [1] "round_na.interp"
#> 
#> 
#> $nbGARCH$datasets
#> [1] "all_total"
#> 
#> $nbGARCH$response
#> $nbGARCH$response$link
#> [1] "negative_binomial"
#> 
#> $nbGARCH$response$type
#> [1] "distribution"
#> 
#> 
#> $nbGARCH$time
#> [1] "year"
#> 
#> 
#> $pGARCH
#> $pGARCH$metadata
#> $pGARCH$metadata$name
#> [1] "pGARCH"
#> 
#> $pGARCH$metadata$descriptor
#> [1] "classic model, requires interpolation"
#> 
#> $pGARCH$metadata$text
#> [1] "pGARCH (Poisson Auto-Regressive Conditional Heteroskedasticity) is a generalized autoregressive conditional heteroskedasticity (GARCH) model with overdispersion (*i.e.*, a Poisson response) fit to the interpolated data at the composite (full site and just control plots) spatial level and both the composite (community) and the articulated (species) ecological levels. The model for each species and the community total is selected and fitted using the `tsglm` function in the **tscount** package [@Liboschik2017a] within the `pGARCH()` function. GARCH models are generalized ARMA models and are defined according to their link function, response distribution, and two model structure parameters -- the number of autoregressive terms (p) and the order of the moving average (q), and are represented as GARCH(p, q) [@Liboschik2017a]. The pGARCH model is fit using the log link and a Poisson response, as well as with p = 1 (first-order autoregression) and q = NULL (no moving average). The `tsglm` function in the **tscount** package [@Liboschik2017a] uses a (conditional) quasi-likelihood based approach to inference and models the overdispersion as an additional parameter in a two-step approach. This two-stage approach has only been minimally evaluated, although preliminary simulation-based studies are promising [@Liboschik2017b]."
#> 
#> 
#> $pGARCH$fun
#> [1] "tsglm"
#> 
#> $pGARCH$args
#> $pGARCH$args$ts
#> [1] "abundance"
#> 
#> $pGARCH$args$model
#> $pGARCH$args$model$past_obs
#> [1] 1
#> 
#> $pGARCH$args$model$past_mean
#> [1] 12
#> 
#> 
#> $pGARCH$args$distr
#> [1] "poisson"
#> 
#> $pGARCH$args$link
#> [1] "log"
#> 
#> 
#> $pGARCH$interpolate
#> $pGARCH$interpolate$needed
#> [1] TRUE
#> 
#> $pGARCH$interpolate$fun
#> [1] "round_na.interp"
#> 
#> 
#> $pGARCH$datasets
#> [1] "all_total"
#> 
#> $pGARCH$response
#> $pGARCH$response$link
#> [1] "poisson"
#> 
#> $pGARCH$response$type
#> [1] "distribution"
#> 
#> 
#> $pGARCH$time
#> [1] "year"
#> 
#>