Prepare Max Counts Data Tables for Forecasting

Generate max count annual datasets using existing functions.

prepare_max_counts(
  name = "all_total",
  main = ".",
  settings = directory_settings(),
  species = "total",
  datafile = "Indicators/max_count_all_total.csv",
  filename = "counts_all_total.csv",
  minyear = 1986,
  maxyear = this_year(),
  level = "all",
  save = TRUE,
  overwrite = TRUE,
  quiet = FALSE,
  verbose = FALSE
)

Arguments

name

character name of the data set.

main

character value of the name of the main component of the directory tree.

settings

list of controls for the directory, with defaults set in directory_settings that should generally not need to be altered.

filename

character value of the file for saving the output.

minyear

Earliest year to include

maxyear

Most recent year to include

level

level at which to summarize (colony, subregion, region, all)

save

logical indicator controlling if the output should be saved out.

overwrite

logical indicator of whether or not the existing files should be updated (most users should leave as TRUE).

quiet

logical indicator if progress messages should be quieted.

verbose

logical indicator if detailed messages should be shown.

Value

data.frame for the specified data set.